requirements.txt
# Phishing Simulation Platform - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Email
aiosmtplib==3.0.1
email-validator==2.1.0
premailer==3.10.0  # Inline CSS for emails

# Database
aiosqlite==0.19.0
libsql-client==0.3.0  # Turso client

# Security
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Analytics & Tracking
pandas==2.1.4
numpy==1.26.3

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CLI
click==8.1.7
rich==13.7.0  # Beautiful terminal output

# Scheduling
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
faker==22.0.0  # Generate test data

# Monitoring
prometheus-client==0.19.0
